requirements.txt
annotated-types
certifi
charset-normalizer

httplib2
idna
pydantic
pydantic_core
pyparsing
python-dotenv
requests
ruff
tqdm
typing-inspection
typing_extensions
urllib3
flask
